Put `list_of` elsewhere and explain it.
